refactor: extract ingredient parser and nutrition data into separate modules
- Move parsing logic (types, constants, regex, matchers, parseIngredients) from RecipeNutrition.tsx into src/services/ingredientParser.ts
- Move lookup constants (MODIFIERS, INGREDIENT_OVERRIDES) from nutritionService.ts into src/services/nutritionData.ts
- Add JSDoc comments to regex building blocks (NUM, RANGE_SEP) and matcher functions
- Simplify ingredient line detection to only accept '-' list markers
- Remove ~80 same-word-only entries from INGREDIENT_OVERRIDES that provided no value over fuzzy search
- Add Zimt and Sojahack to custom_nutrition.json
