+++ src/services/nutritionService.ts
@@ -1,6 +1,8 @@
 import Fuse, { FuseResult } from 'fuse.js'
 import { newStemmer } from 'snowball-stemmers'
 
+import { INGREDIENT_OVERRIDES, MODIFIERS } from './nutritionData'
+
 export interface NutritionEntry {
   id: string
   name: string
@@ -46,240 +48,6 @@ export interface NutritionDetailedResult {
 }
 
 const stemmer = newStemmer('german')
-
-// Words that describe preparation, size, or state but don't identify the ingredient.
-// Common inflected forms are included to avoid a dependency on runtime stemming for filtering.
-const MODIFIERS = new Set([
-  // frozen / processing
-  'tk',
-  'tiefkühl',
-  'tiefgefroren',
-  'tiefgekühlt',
-  // preparation methods
-  'frisch',
-  'frische',
-  'roh',
-  'getrocknet',
-  'getrocknete',
-  'getrockneter',
-  'getrocknetes',
-  'gehackt',
-  'gehackte',
-  'gehackter',
-  'gehacktes',
-  'gemahlen',
-  'gemahlene',
-  'gemahlener',
-  'gemahlenes',
-  'gerieben',
-  'geriebene',
-  'geröstet',
-  'geröstete',
-  'eingelegt',
-  'eingelegte',
-  'geschält',
-  'geschälte',
-  'gekocht',
-  'gekochte',
-  'gegart',
-  'gegarte',
-  'gebacken',
-  'gebackene',
-  // size
-  'klein',
-  'kleine',
-  'kleiner',
-  'kleines',
-  'groß',
-  'große',
-  'großer',
-  'großes',
-  'mittel',
-  'mittlere',
-  'ganz',
-  'ganze',
-  'ganzer',
-  'ganzes',
-  'gute',
-  'fein',
-  'feine',
-  'feiner',
-  'grob',
-  'grobe',
-  'grober',
-  // temperature / state
-  'kalt',
-  'kalte',
-  'kalter',
-  'kaltes',
-  'warm',
-  'warme',
-  'warmer',
-  'heißes',
-  // connectors
-  'oder',
-  'und',
-  'mit',
-])
-
-/**
- * Ingredients with no useful BLS entry — skipped before any search so the
- * fuzzy matcher cannot pick up unrelated compounds (e.g. "zimt" → "Zimtsterne").
- */
-const SKIP_INGREDIENTS = new Set(['zimt']) // cSpell:ignore zimt
-
-/**
- * Maps common German ingredient names (lowercase, modifier-stripped) to a more
- * specific BLS search query. Checked before fuzzy search to avoid ambiguous
- * short queries matching obscure compounds (e.g. flour matching arrowroot flour).
- */
-// cSpell:disable
-const INGREDIENT_OVERRIDES: Record<string, string> = {
-  // Flour & grains
-  mehl: 'Weizen Mehl, Type 550',
-  weizenmehl: 'Weizen Mehl, Type 550',
-  dinkelmehl: 'Dinkel Mehl, Type 630',
-  roggenmehl: 'Roggen Mehl',
-  maismehl: 'Mais Mehl',
-  reismehl: 'Reis Mehl',
-  speisestärke: 'Speisestärke',
-  stärke: 'Speisestärke',
-  maisstärke: 'Speisestärke',
-  haferflocken: 'Hafer Flocken',
-  reis: 'Reis',
-  nudeln: 'Nudeln',
-  pasta: 'Nudeln',
-  spaghetti: 'Spaghetti',
-  semmelbrösel: 'Semmelbrösel',
-  paniermehl: 'Semmelbrösel',
-  // Dairy
-  butter: 'Butter',
-  margarine: 'Margarine',
-  sojamilch: 'Sojadrink',
-  milch: 'Vollmilch',
-  sahne: 'Schlagsahne',
-  schlagsahne: 'Schlagsahne',
-  joghurt: 'Joghurt',
-  quark: 'Speisequark',
-  magerquark: 'Speisequark',
-  frischkäse: 'Frischkäse',
-  schmand: 'Schmand',
-  käse: 'Gouda',
-  parmesan: 'Parmesan',
-  mozzarella: 'Mozzarella',
-  gouda: 'Gouda',
-  edamer: 'Edamer',
-  // Eggs
-  ei: 'Hühnerei',
-  eier: 'Hühnerei',
-  // Oils & fats
-  öl: 'Rapsöl',
-  olivenöl: 'Olivenöl',
-  rapsöl: 'Rapsöl',
-  sonnenblumenöl: 'Sonnenblumenöl',
-  kokosöl: 'Kokosöl',
-  // Sweeteners
-  zucker: 'Zucker weiß',
-  puderzucker: 'Puderzucker',
-  honig: 'Honig',
-  ahornsirup: 'Ahornsirup',
-  // Salt, spices & herbs
-  salz: 'Speisesalz',
-  pfeffer: 'Pfeffer',
-  paprikapulver: 'Paprikapulver',
-  muskat: 'Muskatnuss',
-  kurkuma: 'Kurkuma',
-  oregano: 'Oregano',
-  basilikum: 'Basilikum',
-  petersilie: 'Petersilie',
-  thymian: 'Thymian',
-  rosmarin: 'Rosmarin',
-  majoran: 'Majoran',
-  koriander: 'Koriander',
-  // Leavening & baking
-  backpulver: 'Backpulver',
-  hefe: 'Hefe',
-  natron: 'Natriumhydrogencarbonat',
-  vanille: 'Vanilleextrakt',
-  vanillezucker: 'Vanillezucker',
-  // Vegetables
-  tomate: 'Tomate',
-  tomaten: 'Tomate',
-  zwiebel: 'Speisezwiebel',
-  zwiebeln: 'Speisezwiebel',
-  knoblauch: 'Knoblauch',
-  karotte: 'Karotte',
-  karotten: 'Karotte',
-  möhre: 'Möhre',
-  möhren: 'Möhre',
-  kartoffel: 'Kartoffel',
-  kartoffeln: 'Kartoffel',
-  zucchini: 'Zucchini',
-  paprika: 'Gemüsepaprika',
-  spinat: 'Spinat',
-  brokkoli: 'Broccoli',
-  blumenkohl: 'Blumenkohl',
-  gurke: 'Gurke',
-  champignon: 'Champignon',
-  champignons: 'Champignon',
-  pilze: 'Champignon',
-  lauch: 'Lauch',
-  sellerie: 'Sellerie',
-  tomatenmark: 'Tomatenmark',
-  // Legumes
-  erbsen: 'Erbse grün',
-  bohnen: 'Kidneybohne',
-  linsen: 'Linse',
-  kichererbsen: 'Kichererbse',
-  // Proteins
-  hähnchen: 'Hähnchenbrustfilet',
-  hühnchen: 'Hähnchenbrustfilet',
-  rindfleisch: 'Rindfleisch',
-  schweinefleisch: 'Schweinefleisch',
-  hackfleisch: 'Rinderhackfleisch',
-  lachs: 'Lachs',
-  thunfisch: 'Thunfisch',
-  garnelen: 'Garnelen',
-  tofu: 'Tofu',
-  // Nuts & seeds
-  mandeln: 'Mandel süß',
-  haselnüsse: 'Haselnuss',
-  walnüsse: 'Walnuss',
-  erdnüsse: 'Erdnuss',
-  cashews: 'Cashewnuss',
-  sesam: 'Sesam',
-  leinsamen: 'Leinsamen',
-  kürbiskerne: 'Kürbiskerne',
-  sonnenblumenkerne: 'Sonnenblumenkerne',
-  // Fruits
-  apfel: 'Apfel',
-  äpfel: 'Apfel',
-  banane: 'Banane',
-  bananen: 'Banane',
-  zitrone: 'Zitrone',
-  zitronen: 'Zitrone',
-  orange: 'Orange',
-  orangen: 'Orange',
-  erdbeeren: 'Erdbeere',
-  himbeeren: 'Himbeere',
-  blaubeeren: 'Blaubeere',
-  // Condiments & liquids
-  senf: 'Senf',
-  essig: 'Weinessig',
-  sojasoße: 'Sojasoße',
-  brühe: 'Gemüsebrühe',
-  gemüsebrühe: 'Gemüsebrühe',
-  hühnerbrühe: 'Hühnerbrühe',
-  wein: 'Weißwein',
-  rotwein: 'Rotwein',
-  weißwein: 'Weißwein',
-  // Other
-  schokolade: 'Zartbitterschokolade',
-  kakaopulver: 'Kakaopulver',
-  erdnussbutter: 'Erdnussmus',
-  kokosmilch: 'Kokosmilch',
-}
-// cSpell:enable
 
 const splitCommonSuffixes = (text: string): string => {
   // cSpell:disable-next-line
@@ -452,7 +220,6 @@ class NutritionService {
     if (!this.fuse || !this.data) return null
 
     const strippedQuery = stripModifiers(query).toLowerCase()
-    if (SKIP_INGREDIENTS.has(strippedQuery)) return null
 
     // Custom entries take priority over the BLS database.
     const customResult = this.findInCustom(strippedQuery)
